containers-verus: match production's node layout and bisection lowering
The B+tree descent was +15..+21% slower than production. Three earlier passes
eliminated footprint, node count, and bounds checks, and one of them wrote off
branch prediction. That acquittal was wrong, and the way it was wrong is the
main lesson here.

ROOT CAUSE: the LOWERING of the bisection, not its shape.

Production does not write a bisection -- it calls `slice::partition_point`, which
uses `core::hint::select_unpredictable` internally and therefore compiles its
base update to `cmovbe`. Our hand-written loop already had partition_point's
shape (unconditional `size -= half`, so the trip count is exactly log2(n)) and
already READ branchless:

    base = if lt { mid } else { base };

but LLVM's if-conversion heuristic judged the compare predictable and emitted
`ja`/`jmp` anyway. An arithmetic mask `(a & !m) | (b & m)` gets folded straight
back to the same branch -- `/tmp` probe: byte-identical instruction histograms.
On shuffled keys that compare is a coin flip, so it mispredicted at about half
the levels of every descent.

All four bisections (`bplus.rs`'s leaf_find_ge + find_child, `bplus_search.rs`'s
BinarySearch::find_ge + find_gt) now route the base update through
`bplus_layout::sel_usize`. The bisection TAIL deliberately keeps a plain `if`:
it lowers to `adc` on the compare's own flag, which is what partition_point's
tail compiles to and is cheaper than a forced `cmov`. Routing the tail through
sel_usize was measured and reverted. `Branchless` needs no change -- its
compare-accumulate is branch-free by construction.

Shape alone, with the `if` form, was worth <1%. The `cmov` is the whole effect.

Also in this commit, because they are one layout-parity change and share the
benchmark that validates them:

  - Node byte layout: `#[repr(C, align($align))]`, `count: u8`, explicit
    `_pad: u16`, matching production's `define_node!`. The `_pad` is load-bearing
    for CODEGEN, not correctness: without the field spelled out, LLVM splits
    every arena read/write into a scalar flags/count pair plus a data+link
    memcpy, which measured +27.3%. It is threaded through value_of/into_repr/
    from_repr verbatim rather than pinned to 0, so the Tagged round-trip holds
    for any padding bits and no `wf` clause has to mention it.
  - `arr_get`/`arr_set`: unchecked array access where `i < N` is a VERIFIED
    precondition at every call site (`node_wf` bounds count by leaf_cap/key_cap,
    both <= data_len). ~6 dead bounds checks per descent level.

MEASURED (containers-conformance/examples/onesite_bplus.rs, Layout256, n=100k,
one call site, both orders, best-of-5 after 20 warmups):

    row                before    after
    descent (iter)     +23.5%    -2.0%
    sm descent          +2.2%   -10.8%
    redescent (dup)    +29.0%    +7.1%
    insert shuffled    +27.6%   +12.7%
    insert asc          -0.9%    +0.7%
    from_sorted         +843%    +967%

Pure descent at n=400k, both build orders: +20.9% -> -19.1%.

ONE COST, RECORDED HONESTLY: `cmov` is a small pessimization where the branch
genuinely IS predictable, which is why the two ascending rows moved the wrong
way. Trading ~2 points on ascending for ~14 on shuffled is the right call, but
it is a trade, not a free win. `from_sorted` stays blocked on a real bottom-up
bulk loader; `insert shuffled` and `redescent` are now dominated by recursion
overhead (`insert_rec` holds a by-value 256-byte node live across the recursive
call -- 0x600-byte frames -- where production keeps only `(idx, cp)` in a
`path: [(ArenaIdx, usize); 24]` array). Both are follow-ups, not regressions.

THREE HYPOTHESES DIED HERE, each by measurement rather than inspection, and each
looked decisive first:

  - Node read shape (+-1%). Production's `Node` IS its `Repr` (`type Repr =
    $name`), so `from_repr` is `*r` plus a flag mask: one aligned memcpy(256).
    Ours has `is_leaf: bool` where the repr has `flags: u8`, so it rebuilds
    field-by-field: 4 scalar loads + memcpy(248) from unaligned base+4 + 4
    scalar stores. Visible in the disassembly and completely free -- a probe
    measured rebuild-vs-identity at +1.8% / +1.1% / -0.6% across 32KB/1MB/64MB.
    LLVM forwards neither copy.
  - Arena node ordering. Killed by building the same key set ascending (both
    arms take the append fast path and split left-to-right, forcing identical
    ordering) versus shuffled, and timing the same probe: +15.3% vs +15.5%.
  - Footprint (1.000x). Counting global allocator: prod 1048640 bytes / 4096
    nodes / 24.4 keys-per-node vs verus 1048576 / 4096 / 24.4.

THE PROCESS LESSON, documented in doc/design/11-layout-parity.md because an
earlier pass got it wrong in exactly this shape: WHEN A SOURCE-LEVEL
OPTIMIZATION MEASURES AS NO-CHANGE, VERIFY IT WAS COMPILED BEFORE CONCLUDING THE
MECHANISM IS ABSENT. A null result has two readings -- the mechanism is
innocent, or the change never reached the machine code -- and only disassembly
separates them. Re-running the benchmark cannot. The branchless rewrite was
written, verified, benchmarked, measured at +-1%, and used to acquit branch
prediction; the source was branchless and the object code never was. Diffing the
two arms' asm for the SAME function (`cmovbe` at prod's db11, `ja` at ours'
d94f) is what found it.

TRUST. Three new `external_body` markers in bplus_layout (21 -> 24 default,
26 -> 29 with `literal-types`), all documented in doc/design/02-trust-boundary.md.
None widens the assumed-fact inventory exploitably: arr_get/arr_set take
`get_unchecked`'s own documented contract, where `N` is the array's own
const-generic length so no arithmetic relates bound to index; `sel_usize`
contains NO `unsafe` at all and its postcondition IS `if c { b } else { a }` --
`select_unpredictable` is a codegen hint, not a semantic one. All three are also
checked at runtime by tests/external_body_contract_fuzz.rs (arr_get/arr_set
against checked indexing over 500 rounds x 62 slots; sel_usize against if/else
on both booleans over the type's extremes plus 500 random pairs).

HARNESS. examples/onesite_bplus.rs is added here because the perf claims above
are not readable without it. `bulkload.rs` builds prod then verus in a fixed
order inside one iteration, which is exactly the shape Ch. 11 documents as worth
+18% to whichever arm runs second (heap reuse) plus another +18% from hot-loop
alignment -- it cannot distinguish a real 10-30% gap from its own measurement
order. onesite selects the impl at RUNTIME inside one `#[inline(never)]` call
site, times each arm first and last, and takes the best of 5 after 20 warmups.
It is also what showed the descent gap that bulkload's 0.9-1.0x had hidden.

VERIFICATION: unchanged counts, zero proof debt. bplus 139, bplus_layout 311,
bplus_search 9, bplus_tree 124; whole crate 1405 verified / 0 errors. 83 test
binaries pass, 0 failures. clippy and `cargo fmt --check` clean.

The three clippy `if_same_then_else` warnings this introduced in bplus_search
were real: erased proof blocks make two exec branches look identical. Fixed by
nesting each case split inside `proof { }`, matching bplus.rs's existing style.

+++ containers-verus/tests/bplus_contract_fuzz.rs
@@ -77,7 +77,7 @@ macro_rules! fuzz_layout {
 
             // --- shadow views recomputed from the node's PUBLIC fields ---
             fn shadow_keys(n: &$Node) -> Vec<$W> {
-                (0..n.count).map(|i| n.data[i]).collect()
+                (0..(n.count as usize)).map(|i| n.data[i]).collect()
             }
             // child_view as a u128 so u32/u64 compare uniformly with the
             // exec result widened the same way.
@@ -105,7 +105,7 @@ macro_rules! fuzz_layout {
                 let link = lcg.next() as usize; // last child / arena id
                 let mut n: $Node = L::new_leaf();
                 n.is_leaf = false;
-                n.count = count;
+                n.count = (count) as u8;
                 n.data = data;
                 n.link = link as _;
                 n
@@ -122,7 +122,7 @@ macro_rules! fuzz_layout {
                 }
                 let mut n: $Node = L::new_leaf();
                 n.is_leaf = true;
-                n.count = count;
+                n.count = (count) as u8;
                 n.data = data;
                 n
             }
@@ -140,7 +140,7 @@ macro_rules! fuzz_layout {
 
                     // is_leaf / count
                     assert_eq!(L::is_leaf(&n), n.is_leaf, "is_leaf mismatch");
-                    assert_eq!(L::count(&n), n.count, "count mismatch");
+                    assert_eq!(L::count(&n), (n.count as usize), "count mismatch");
                     // link_view
                     assert_eq!(
                         L::link(&n).as_usize() as u128,
@@ -150,13 +150,13 @@ macro_rules! fuzz_layout {
 
                     // key(i) == keys_view[i]  (the data[0..count] projection)
                     let keys = shadow_keys(&n);
-                    for i in 0..n.count {
+                    for i in 0..(n.count as usize) {
                         assert_eq!(L::key(&n, i), keys[i], "key[{i}] mismatch");
                     }
                     // child(i) == child_view(i) for internal nodes, i in 0..=count.
                     // THIS is the u64->usize cast under test for the u64 layouts.
                     if !n.is_leaf {
-                        for i in 0..=n.count {
+                        for i in 0..=(n.count as usize) {
                             let got = L::child(&n, i).as_usize() as u128;
                             assert_eq!(
                                 got,
@@ -208,10 +208,10 @@ macro_rules! fuzz_layout {
                 for _ in 0..20_000 {
                     let mut n = rand_internal(&mut lcg);
                     if n.count == 0 {
-                        n.count = 1;
+                        n.count = (1) as u8;
                         n.data[0] = word(&mut lcg);
                     } // need >= 1 child slot
-                    let i = lcg.upto(n.count + 1); // 0..=count
+                    let i = lcg.upto((n.count as usize) + 1); // 0..=count
                     let v = lcg.next() as usize;
                     let v_idx = <<L as NodeLayout>::ArenaIdx as IndexLike>::try_from_usize(v)
                         .unwrap_or(<<L as NodeLayout>::ArenaIdx as IndexLike>::max());
@@ -229,11 +229,11 @@ macro_rules! fuzz_layout {
                     );
                     assert_eq!(
                         L::count(&n),
-                        before.count,
+                        (before.count as usize),
                         "set_internal_child changed count"
                     );
                     // keys_view unchanged
-                    for j in 0..before.count {
+                    for j in 0..(before.count as usize) {
                         assert_eq!(
                             L::key(&n, j),
                             L::key(&before, j),
@@ -245,7 +245,7 @@ macro_rules! fuzz_layout {
                         vw,
                         "set_internal_child target child[{i}] != v"
                     );
-                    for j in 0..=before.count {
+                    for j in 0..=(before.count as usize) {
                         if j != i {
                             assert_eq!(
                                 L::child(&n, j).as_usize(),
@@ -269,8 +269,8 @@ macro_rules! fuzz_layout {
                     L::set_link(&mut n, v_idx);
                     // ensures: only the link changes.
                     assert_eq!(L::is_leaf(&n), L::is_leaf(&before));
-                    assert_eq!(L::count(&n), before.count);
-                    for j in 0..before.count {
+                    assert_eq!(L::count(&n), (before.count as usize));
+                    for j in 0..(before.count as usize) {
                         assert_eq!(
                             L::key(&n, j),
                             L::key(&before, j),
@@ -290,10 +290,10 @@ macro_rules! fuzz_layout {
                 let mut lcg = Lcg::new(0x1EAF_1227 ^ (KEY_CAP as u64));
                 for _ in 0..20_000 {
                     let mut n = rand_leaf(&mut lcg, false);
-                    if n.count >= LEAF_CAP {
+                    if (n.count as usize) >= LEAF_CAP {
                         continue;
                     } // requires count < leaf_cap
-                    let pos = lcg.upto(n.count + 1); // 0..=count
+                    let pos = lcg.upto((n.count as usize) + 1); // 0..=count
                     let w = word(&mut lcg);
                     let before_keys = shadow_keys(&n);
                     let old_link = n.link;
@@ -315,7 +315,7 @@ macro_rules! fuzz_layout {
                 let mid = L::split_mid();
                 for _ in 0..20_000 {
                     let n = rand_leaf(&mut lcg, true); // FULL leaf (count == leaf_cap)
-                    assert_eq!(n.count, LEAF_CAP);
+                    assert_eq!((n.count as usize), LEAF_CAP);
                     let pos = lcg.upto(LEAF_CAP + 1); // 0..=leaf_cap
                     let w = word(&mut lcg);
                     // combined = keys_view.insert(pos, w)
@@ -347,7 +347,7 @@ macro_rules! fuzz_layout {
                 for _ in 0..20_000 {
                     // FULL internal node: count == key_cap.
                     let mut n = rand_internal(&mut lcg);
-                    n.count = KEY_CAP;
+                    n.count = (KEY_CAP) as u8;
                     for i in 0..KEY_CAP {
                         n.data[i] = word(&mut lcg);
                     }
@@ -426,19 +426,23 @@ macro_rules! fuzz_layout {
                 let mut lcg = Lcg::new(0x142E_0001 ^ (LEAF_CAP as u64));
                 for _ in 0..20_000 {
                     let mut n = rand_internal(&mut lcg);
-                    if n.count >= KEY_CAP {
+                    if (n.count as usize) >= KEY_CAP {
                         continue;
                     } // requires count < key_cap
-                    let pos = lcg.upto(n.count + 1);
+                    let pos = lcg.upto((n.count as usize) + 1);
                     let w = word(&mut lcg);
-                    let before = (0..n.count).map(|i| n.data[i]).collect::<Vec<$W>>();
+                    let before = (0..(n.count as usize))
+                        .map(|i| n.data[i])
+                        .collect::<Vec<$W>>();
                     L::internal_key_insert(&mut n, pos, w);
                     // ensures: count+1, keys == before.insert(pos, w), still internal.
                     assert!(!L::is_leaf(&n));
                     assert_eq!(L::count(&n), before.len() + 1, "internal_key_insert count");
                     let mut want = before.clone();
                     want.insert(pos, w);
-                    let got = (0..n.count).map(|i| n.data[i]).collect::<Vec<$W>>();
+                    let got = (0..(n.count as usize))
+                        .map(|i| n.data[i])
+                        .collect::<Vec<$W>>();
                     assert_eq!(got, want, "internal_key_insert keys pos={pos}");
                 }
             }
